Change in the configuration of mission variables files
