+++ lib/utils/mixins/mission_variables/class_mission_variable.dart
@@ -1,0 +1,29 @@
+import 'package:zenith_monitor/utils/helpers/string_to_pattern.dart';
+
+class MissionVariable<T> {
+  late String _variableName;
+  late String _variableType;
+  T? _variableValue;
+
+  MissionVariable(String variableName, String variableType) {
+    _variableName = variableName;
+    _variableType = stringToPattern(variableType);
+    _variableValue = null;
+  }
+
+  String getVariableName() {
+    return _variableName;
+  }
+
+  String getVariableType() {
+    return _variableType;
+  }
+
+  T? getVariableValue() {
+    return _variableValue;
+  }
+
+  void addValue(T value) {
+    _variableValue = value;
+  }
+}
